Move util functions back to processor file
