+++ snuba/datasets/events_processor.py
@@ -1,20 +1,21 @@
 from datetime import datetime, timedelta
-from hashlib import md5
 import logging
-import re
-import simplejson as json
 import six
 import _strptime  # NOQA fixes _strptime deferred import issue
 
 from snuba import settings
-from snuba.util import force_bytes
-from snuba.processor import MessageProcessor
+from snuba.processor import (
+    _as_dict_safe,
+    _boolify,
+    _collapse_uint32,
+    _ensure_valid_date,
+    _floatify,
+    _hashify,
+    _unicodify,
+    MessageProcessor,
+)
 
 logger = logging.getLogger('snuba.processor')
-
-
-HASH_RE = re.compile(r'^[0-9a-f]{32}$', re.IGNORECASE)
-MAX_UINT32 = 2 ** 32 - 1
 
 
 class InvalidMessageType(Exception):
@@ -374,80 +375,3 @@ class EventsProcessor(MessageProcessor):
 
 class EventTooOld(Exception):
     pass
-
-
-def _as_dict_safe(value):
-    if value is None:
-        return {}
-    if isinstance(value, dict):
-        return value
-    rv = {}
-    for item in value:
-        if item is not None:
-            rv[item[0]] = item[1]
-    return rv
-
-
-def _collapse_uint32(n):
-    if (n is None) or (n < 0) or (n > MAX_UINT32):
-        return None
-    return n
-
-
-def _boolify(s):
-    if s is None:
-        return None
-
-    if isinstance(s, bool):
-        return s
-
-    s = _unicodify(s)
-
-    if s in ('yes', 'true', '1'):
-        return True
-    elif s in ('false', 'no', '0'):
-        return False
-
-    return None
-
-
-def _floatify(s):
-    if not s:
-        return None
-
-    if isinstance(s, float):
-        return s
-
-    try:
-        s = float(s)
-    except (ValueError, TypeError):
-        return None
-    else:
-        return s
-
-    return None
-
-
-def _unicodify(s):
-    if s is None:
-        return None
-
-    if isinstance(s, dict) or isinstance(s, list):
-        return json.dumps(s)
-
-    return six.text_type(s)
-
-
-def _hashify(h):
-    if HASH_RE.match(h):
-        return h
-    return md5(force_bytes(h)).hexdigest()
-
-
-def _ensure_valid_date(dt):
-    if dt is None:
-        return None
-    seconds = (dt - datetime(1970, 1, 1)).total_seconds()
-    if _collapse_uint32(seconds) is None:
-        return None
-    return dt
