+++ snuba/replacer.py
@@ -9,7 +9,8 @@ from batching_kafka_consumer import AbstractBatchWorker
 
 from . import settings
 from snuba.clickhouse import escape_col
-from snuba.datasets.events_processor import _hashify, InvalidMessageType, InvalidMessageVersion
+from snuba.processor import _hashify
+from snuba.datasets.events_processor import InvalidMessageType, InvalidMessageVersion
 from snuba.redis import redis_client
 from snuba.util import escape_string
